split endpoint into smaller files
