+++ src/rest/mod.rs
@@ -1,5 +1,10 @@
 mod directive;
 mod endpoint;
 mod endpoint_set;
+mod partial_request;
+mod path;
+mod query_params;
 mod type_map;
+mod typed_variables;
+
 pub use endpoint_set::EndpointSet;
